Module aws_sdk_managedblockchain::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

CreateMember

Fluent builder constructing a request to CreateMember.

CreateNetwork

Fluent builder constructing a request to CreateNetwork.

CreateNode

Fluent builder constructing a request to CreateNode.

CreateProposal

Fluent builder constructing a request to CreateProposal.

DeleteMember

Fluent builder constructing a request to DeleteMember.

DeleteNode

Fluent builder constructing a request to DeleteNode.

GetMember

Fluent builder constructing a request to GetMember.

GetNetwork

Fluent builder constructing a request to GetNetwork.

GetNode

Fluent builder constructing a request to GetNode.

GetProposal

Fluent builder constructing a request to GetProposal.

ListInvitations

Fluent builder constructing a request to ListInvitations.

ListMembers

Fluent builder constructing a request to ListMembers.

ListNetworks

Fluent builder constructing a request to ListNetworks.

ListNodes

Fluent builder constructing a request to ListNodes.

ListProposalVotes

Fluent builder constructing a request to ListProposalVotes.

ListProposals

Fluent builder constructing a request to ListProposals.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

RejectInvitation

Fluent builder constructing a request to RejectInvitation.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateMember

Fluent builder constructing a request to UpdateMember.

UpdateNode

Fluent builder constructing a request to UpdateNode.

VoteOnProposal

Fluent builder constructing a request to VoteOnProposal.