Module aws_sdk_managedblockchain::error

Modules

access_denied_exception

See AccessDeniedException

illegal_action_exception

See IllegalActionException

internal_service_error_exception

See InternalServiceErrorException

invalid_request_exception

See InvalidRequestException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_limit_exceeded_exception

See ResourceLimitExceededException

resource_not_found_exception

See ResourceNotFoundException

resource_not_ready_exception

See ResourceNotReadyException

throttling_exception

See ThrottlingException

too_many_tags_exception

See TooManyTagsException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

CreateMemberError

CreateNetworkError

CreateNodeError

CreateProposalError

DeleteMemberError

DeleteNodeError

GetMemberError

GetNetworkError

GetNodeError

GetProposalError

IllegalActionException

InternalServiceErrorException

The request processing has failed because of an unknown error, exception or failure.

InvalidRequestException

The action or operation requested is invalid. Verify that the action is typed correctly.

ListInvitationsError

ListMembersError

ListNetworksError

ListNodesError

ListProposalVotesError

ListProposalsError

ListTagsForResourceError

RejectInvitationError

ResourceAlreadyExistsException

A resource request is issued for a resource that already exists.

ResourceLimitExceededException

The maximum number of resources of that type already exist. Ensure the resources requested are within the boundaries of the service edition and your account limits.

ResourceNotFoundException

A requested resource does not exist. It may have been deleted or referenced inaccurately.

ResourceNotReadyException

The requested resource exists but is not in a status that can complete the operation.

TagResourceError

ThrottlingException

The request or operation could not be performed because a service is throttling requests. The most common source of throttling errors is launching EC2 instances such that your service limit for EC2 instances is exceeded. Request a limit increase or delete unused resources if possible.

TooManyTagsException

UntagResourceError

UpdateMemberError

UpdateNodeError

VoteOnProposalError

Enums

CreateMemberErrorKind

CreateNetworkErrorKind

CreateNodeErrorKind

CreateProposalErrorKind

DeleteMemberErrorKind

DeleteNodeErrorKind

GetMemberErrorKind

GetNetworkErrorKind

GetNodeErrorKind

GetProposalErrorKind

ListInvitationsErrorKind

ListMembersErrorKind

ListNetworksErrorKind

ListNodesErrorKind

ListProposalVotesErrorKind

ListProposalsErrorKind

ListTagsForResourceErrorKind

RejectInvitationErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateMemberErrorKind

UpdateNodeErrorKind

VoteOnProposalErrorKind