Module aws_sdk_managedblockchain::model

Modules

approval_threshold_policy

See ApprovalThresholdPolicy

invitation

See Invitation

invite_action

See InviteAction

log_configuration

See LogConfiguration

log_configurations

See LogConfigurations

member

See Member

member_configuration

See MemberConfiguration

member_fabric_attributes

See MemberFabricAttributes

member_fabric_configuration

See MemberFabricConfiguration

member_fabric_log_publishing_configuration

See MemberFabricLogPublishingConfiguration

member_framework_attributes

See MemberFrameworkAttributes

member_framework_configuration

See MemberFrameworkConfiguration

member_log_publishing_configuration

See MemberLogPublishingConfiguration

member_summary

See MemberSummary

network

See Network

network_ethereum_attributes

See NetworkEthereumAttributes

network_fabric_attributes

See NetworkFabricAttributes

network_fabric_configuration

See NetworkFabricConfiguration

network_framework_attributes

See NetworkFrameworkAttributes

network_framework_configuration

See NetworkFrameworkConfiguration

network_summary

See NetworkSummary

node

See Node

node_configuration

See NodeConfiguration

node_ethereum_attributes

See NodeEthereumAttributes

node_fabric_attributes

See NodeFabricAttributes

node_fabric_log_publishing_configuration

See NodeFabricLogPublishingConfiguration

node_framework_attributes

See NodeFrameworkAttributes

node_log_publishing_configuration

See NodeLogPublishingConfiguration

node_summary

See NodeSummary

proposal

See Proposal

proposal_actions

See ProposalActions

proposal_summary

See ProposalSummary

remove_action

See RemoveAction

vote_summary

See VoteSummary

voting_policy

See VotingPolicy

Structs

ApprovalThresholdPolicy

A policy type that defines the voting rules for the network. The rules decide if a proposal is approved. Approval may be based on criteria such as the percentage of YES votes and the duration of the proposal. The policy applies to all proposals and is specified when the network is created.

Invitation

An invitation to an AWS account to create a member and join the network.

InviteAction

An action to invite a specific AWS account to create a member and join the network. The InviteAction is carried out when a Proposal is APPROVED.

LogConfiguration

A configuration for logging events.

LogConfigurations

A collection of log configurations.

Member

Member configuration properties.

MemberConfiguration

Configuration properties of the member.

MemberFabricAttributes

Attributes of Hyperledger Fabric for a member in a Managed Blockchain network using the Hyperledger Fabric framework.

MemberFabricConfiguration

Configuration properties for Hyperledger Fabric for a member in a Managed Blockchain network using the Hyperledger Fabric framework.

MemberFabricLogPublishingConfiguration

Configuration properties for logging events associated with a member of a Managed Blockchain network using the Hyperledger Fabric framework.

MemberFrameworkAttributes

Attributes relevant to a member for the blockchain framework that the Managed Blockchain network uses.

MemberFrameworkConfiguration

Configuration properties relevant to a member for the blockchain framework that the Managed Blockchain network uses.

MemberLogPublishingConfiguration

Configuration properties for logging events associated with a member of a Managed Blockchain network.

MemberSummary

A summary of configuration properties for a member.

Network

Network configuration properties.

NetworkEthereumAttributes

Attributes of Ethereum for a network.

NetworkFabricAttributes

Attributes of Hyperledger Fabric for a network.

NetworkFabricConfiguration

Hyperledger Fabric configuration properties for the network.

NetworkFrameworkAttributes

Attributes relevant to the network for the blockchain framework that the network uses.

NetworkFrameworkConfiguration

NetworkSummary

A summary of network configuration properties.

Node

Configuration properties of a node.

NodeConfiguration

Configuration properties of a node.

NodeEthereumAttributes

Attributes of an Ethereum node.

NodeFabricAttributes

Attributes of Hyperledger Fabric for a peer node on a Hyperledger Fabric network on Managed Blockchain.

NodeFabricLogPublishingConfiguration

Configuration properties for logging events associated with a peer node owned by a member in a Managed Blockchain network.

NodeFrameworkAttributes

Attributes relevant to a node on a Managed Blockchain network for the blockchain framework that the network uses.

NodeLogPublishingConfiguration

Configuration properties for logging events associated with a peer node on a Hyperledger Fabric network on Managed Blockchain.

NodeSummary

A summary of configuration properties for a node.

Proposal

Properties of a proposal on a Managed Blockchain network.

ProposalActions

ProposalSummary

Properties of a proposal.

RemoveAction

An action to remove a member from a Managed Blockchain network as the result of a removal proposal that is APPROVED. The member and all associated resources are deleted from the network.

VoteSummary

VotingPolicy

Enums

Edition

Framework

InvitationStatus

MemberStatus

NetworkStatus

NodeStatus

ProposalStatus

StateDbType

ThresholdComparator

VoteValue