Module aws_sdk_managedblockchain::operation

Structs

CreateMember

Creates a member within a Managed Blockchain network.

CreateNetwork

Creates a new blockchain network using Amazon Managed Blockchain.

CreateNode

Creates a node on the specified blockchain network.

CreateProposal

Creates a proposal for a change to the network that other members of the network can vote on, for example, a proposal to add a new member to the network. Any member can create a proposal.

DeleteMember

Deletes a member. Deleting a member removes the member and all associated resources from the network. DeleteMember can only be called for a specified MemberId if the principal performing the action is associated with the AWS account that owns the member. In all other cases, the DeleteMember action is carried out as the result of an approved proposal to remove a member. If MemberId is the last member in a network specified by the last AWS account, the network is deleted also.

DeleteNode

Deletes a node that your AWS account owns. All data on the node is lost and cannot be recovered.

GetMember

Returns detailed information about a member.

GetNetwork

Returns detailed information about a network.

GetNode

Returns detailed information about a node.

GetProposal

Returns detailed information about a proposal.

ListInvitations

Returns a list of all invitations for the current AWS account.

ListMembers

Returns a list of the members in a network and properties of their configurations.

ListNetworks

Returns information about the networks in which the current AWS account participates.

ListNodes

Returns information about the nodes within a network.

ListProposalVotes

Returns the list of votes for a specified proposal, including the value of each vote and the unique identifier of the member that cast the vote.

ListProposals

Returns a list of proposals for the network.

ListTagsForResource

Returns a list of tags for the specified resource. Each tag consists of a key and optional value.

RejectInvitation

Rejects an invitation to join a network. This action can be called by a principal in an AWS account that has received an invitation to create a member and join a network.

TagResource

Adds or overwrites the specified tags for the specified Amazon Managed Blockchain resource. Each tag consists of a key and optional value.

UntagResource

Removes the specified tags from the Amazon Managed Blockchain resource.

UpdateMember

Updates a member configuration with new parameters.

UpdateNode

Updates a node configuration with new parameters.

VoteOnProposal

Casts a vote for a specified ProposalId on behalf of a member. The member to vote as, specified by VoterMemberId, must be in the same AWS account as the principal that calls the action.